1。解决了一些小BUG，并将库存项的入库克重属性去掉了

+++ Source/LJH.Inventory.UI/Forms/Inventory/Frm小件库存管理.cs
@@ -196,7 +196,7 @@ namespace LJH.Inventory.UI.Forms.Inventory
             if(sr.Product.Length .HasValue ) row.Cells["colLength"].Value = sr.Product.Length.Value.Trim();
             row.Cells["colRealThick"].Value = sr.Real克重;
             row.Cells["colInventoryDate"].Value = sr.AddDate.ToString("yyyy-MM-dd");
-            row.Cells["colCount"].Value = sr.Count;
+            row.Cells["colCount"].Value = (int)sr.Count;
             row.Cells["colDeliverySheet"].Value = sr.DeliverySheet;
             row.Cells["colCustomer"].Value = sr.Customer;
             if (_AllSuppliers != null)
